// src/features/iphone-3d-overview/types/params.types.ts
export interface IphoneSizeType {
  label: string;
  value: "small" | "large";
}

export interface IphoneModelType {
  id: number;
  title: string;
  color: string[];
  material: string;
  modelPath?: string;
}

export interface IphoneColorType {
  id: number;
  title: string;
  imageSrc: string;
  color: string | string[];
}